requirements.txt
# Core LangChain framework
langchain
langchain-core

# Provider-specific libraries
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-community # For Ollama

# For loading environment variables from .env file
python-dotenv


# Libraries for file parsing in wordlist generator
PyPDF2
python-docx

fastapi
uvicorn
jinja2
pydantic
requests
python-multipart
